requirements_gemini.txt
google-generativeai>=0.8.0
pillow>=10.0.0
pydantic>=2.0.0
loguru>=0.7.0 